fixed bug on GETone, added 404 for DELETE, added CORS our project, added test index.html with fetch request

+++ src/modules/todo/controllers/todo.controller.ts
@@ -63,7 +63,14 @@ export class TodoController {
   }
 
   @Delete(':id')
-  deleteAction(@Param('id') id: string): Promise<void> {
-    return this.todoService.remove(id);
+  async deleteAction(@Param('id') id: string): Promise<{ success: boolean }> {
+    const todo = await this.todoService.findOne(id);
+    if (todo === undefined) {
+      throw new NotFoundException(`Todo with id=${id} not found`);
+    }
+    await this.todoService.remove(id);
+    return {
+      success: true,
+    };
   }
 }
